socket-utils: Ensure child processes are terminated upon socket close
Addresses issue #7.

Added some rudimentary PID to FD tracking so that when we close a file
descriptor opened by `open_serial_socket()`, the process behind that socket
(if any) is explicitly terminated.

This unfortunately requires us to add a new function called
`close_serial_socket()`, but in this particular case the change was not
onerous.

We are only handling five such sockets for now. If there are ever more than
that many system sockets opened through this mechanism, then something is
likely seriously wrong.

The socket tracking code is super simple and possibly not thread safe,
which is fine in this case since these methods will only ever be called
from the same thread.

I really wish these sorts of antics were not necessary, but all other
methods of handling this sort of thing seem to come up short.