NCPInstanceBase: Detect and report invalid state transitions
This change was deemed necessary because we saw a case where a
device entered the `OFFLINE` state from the `UPGRADING` state, 
which caused all sorts of havoc.

This change adds some additional checks when a state transition
is triggered so that we detect and report invalid state transitions.
In some cases, invalid state transitions are actually outright
prevented, such as transitioning directly from a detached state
(like `FAULT` or `UPGRADING`) to any state other than `UNINITIALIZED`.

All invalid state transitions are clearly indicated to be bugs
in the logs.