SpinelNCPControlInterface: Fix config_gateway, add/remove external route
This commit changes how the `config_gateway`, `add/remove_external_route`
are implemented in `SpinelNCPControlInterface` (spinel plugin). It ensures
that after an update to any local network data (mesh prefixes or external
routes), the new data is registered with the leader. This is done by adding a
new Spinel property `THREAD_ALLOW_LOCAL_NET_DATA_CHANGE` with boolean type.
This property is set to 'true' before modifying any properties related to
local network data. When done, this property is set to 'false' causing the
network data to be registered with leader. This commit also adds a new
SpinelNCPTask, namely `SpinelNCPTaskChangeNetData` which implements the
common logic for changing any net data property.

This commit is related to openthread issue #272.