wpan-connman-plugin: Ensure that network info is retained in case of untimely Connman restarts
This commit changes how the disconnect is handled in the
wpan-connman-plugin. If a disconnect command/call is received
and if this is not user initiated, the current NCP state is checked:
If it is in `ASSOCIATING` or `NCP_STATE_CREDENTIALS_NEEDED` states,
instead of issuing a `Leave` command to wpantund, we `Reset` the NCP.
This ensures that ongoing association process is stopped, but will
not remove/erase any retained NCP network info.