Adding NetworkRetain class and wpanretain script
This commit adds a new class `NetworkRetain` to wpantund. This class
provides hooks to enable retaining (saving/recalling) network info.
This is realized by adding a new configuration property to wpantund
`Config:Daemon:NetworkRetainCommand`. This property provides the
command to perform the network retain operations. The `NetworkRetain`
class executes the retain command from a forked child process to
request  a save operation (when NCP state transition to "associated" )
or an erase (upon a `leave`) or network info recall and restore (upon
NCP reset or a wpantund restart). The `NetworkRetain` uses the same
model as in `FirmwareUpgrade` module (i.e., the forked process
is started before wpatund possibly dropping privileges).

This commit also adds a script `wpanretain.sh` which can be used as
the network retain command. This script implements the save, recall
and erase operations. It uses `wpanctl get property` commands and
parses the responses to collect network parameters to save in a
config file.