ncp_spinel: Tone down the wording on a common error message.
At initialization, we have a list of properties we want to
try and fetch so that we can make sure that the state in
`wpantund` matches that of the NCP. However, not all of these
properties might be implemented, which is OK. In such a case
we log a warning and continue.

The wording of this error message has caused some concern, so
I've reworded the error message and also change the warning
to also lookup the textual representation of both the property
and the error, so that it is clear what exactly happened.