NCPInstanceBase: Add `is_initializing_ncp()` and `set_initializing_ncp()`
These functions are used to indicate when the NCP driver plugin
is in the process of initializing the NCP. When the NCP is
"initializing", state change events are withheld and attempts to
query `NCP:State` will indicate that the NCP is uninitialized.
Once `set_initializing_ncp(false)` is called, the state change
signal is emitted to whatever state was set while initializing.