ncp: Stop printing out `get_property` debug prints during initialization.
During initialization, logs were being polluted with lines like:

    wpantund[12345] <info> get_property: key: "NCP:State"
    wpantund[12345] <info> get_property: key: "NCP:State"
    wpantund[12345] <info> get_property: key: "NCP:State"

These lines are useless, and come from the the main wpantund run-loop
attempting to determine if the NCP instance is ready to be announced
via D-Bus. However, this log line is still generally useful, this
commit adjusts the behavior to only print this message when we are not
initializing.