ncp-spinel: Remove obsolete `start_new_task()` design pattern
The original design pattern for `start_new_task()` was that if
the method returned false then the callback was not called and
would need to be manually dispatched. However, this method has
not behaved in this way for quite some time---the callback is
always be dispatched. This means that a lot of code that calls
this method is broken: the callback can be called twice under
some specific scenarios, leading to a crash.

This change eliminates the return value from the method and
removes all checks for the return value from all instances where
this method is called.