wpantund: Support for raw packet capture (Packet sniffing)
This commit adds support for the capture and logging of raw packets
from the radio, assuming the underlying NCP supports it.

Note that this feature captures raw packets from the radio, which will
currently always be 802.15.4 frames (but that may not always be the
case in the future). If you just want to sniff IP traffic, you can
simply use Wireshark or tcpdump directly.

Packets are captured in the PCAP format, which is the same format used
by Wireshark and tcpdump. You can start a packet capture using the
`pcap` command in `wpanctl`:

    Syntax:
       pcap [args] <capture-file>
    Options:
       -h/--help                 Print Help
       -t/--timeout              Set timeout period [ms]
       -f                        Force use of stdout, even if it is a tty

If you to not specify a file, `stdout` is used. If `stdout` is a TTY,
it will refuse unless you also specify `-f`.

For real-time monitoring in Wireshark, you can capture to a named pipe
(created using `mkfifo`) and have Wireshark capture from that.

Note that support for promiscuous sniffing is not yet supported. That
feature will be added in a subsequent pull request.

Raw packet capture is being added to OpenThread NCPs via [PR #644][1].

[1]: https://github.com/openthread/openthread/pull/644