NCPInstanceBase: Make sure we clear the global address table upon reset
Upon leaving a network, we were not clearing `wpantund`'s global address
table. This would cause these old addresses to be re-added to the interface
when the same instance of `wpantund` associated with a new network.

This change fixes this issue by making sure that we clear the global
address table whenever we reset the interface.