ncp-spinel: Add support for promiscuous packet capture
This change allows the `wpanctl pcap` command to work when in an
`offline` state. When started while `offline`, it puts the NCP into a
promiscuous sniffing mode.

If you subsequently associate with a network while sniffing
promiscuously, promiscuous mode is turned off and behavior reverts to
the same as if you had started `wpanctl pcap` while associated.

This change also restores packet capture state properly after a NCP
reset.