Add support to save and restore certain radio properties during initialization
This commit adds support in `SpinelNcpInstance` to save and restore
certain settings/properties. This is currently used to remember the
jamming detection related parameters.

The settings are stored in a hash table which maps the property name
(as `std::string`) to a `SettingsEntry` struct.  The `SettingsEntry`
contains the spinel command to set the value of the property, along
with an optional capability value. At the end of NCP initialization, if
the `capability` value is given and it is present in the list of NCP
capabilities, the associated spinel command is sent to NCP to set/
restore the value of the the property.