Ensure the event-handler gets the original spinel data message
This commit changes the `handle_ncp_spinel_value_is()` in
`SpinelNcpInstance` to save a copy of original spinel `data_ptr` and
`dat_len` (before parsing the spinel message). The saved copy of
`data_ptr` and `data_len` is then passed at the end to any
event-handler in `process_event()` call.