NCPTaskSendCommand: Add support for `ReplyUnpacker` function pointer
This commit modifies the `SpinelNCPTaskSendCommand` to add support
for handling user-specified reply formats.

The `Factory` class currently allows the caller to provide a simple
single type spinel reply format. This commit adds support in this
class to alternatively specify a `ReplyUnpacker` function pointer.
When specified, the `ReplyUnpacker` function is invoked to
unpack/decode the reply message.