Do not use stdin or stdout as an l-value
In a handful of places, `stdin` and `stdout` were being used as
l-values. While `stdin` and `stdout` can be used as an l-value
in many standard library implementations, it is explicitly stated
in C89 that it is acceptable for implementations to not make
these symbols l-values. Issue #89 is an example of this.

This change rewrites all places where we are using these variables
as l-values so that `wpantund` can be used on those platforms.

This commit fixes and closes #89.