build: Include a stub copy of boost for systems that don't have it. (#129)
We've run into trouble with the dependency on
[Boost](http://boost.org) recently, and since the parts we are using
are entirely implemented as template headers we figured the easiest
path forward was to simply include those headers with wpantund.

On systems which already have Boost installed, that version will be
used. The bundled version will only be used when no usable copy of
Boost can be found.