Make form/join/resume/leave less fragile.
One of the core cornerstones of Spinel is to not treat the failure of
largely unrelated commands to be fatal. Currently, the absence of the
net save commands is going to cause forming, joining, resuming, and
leaving to fail. This is entirely unnecessary, since the failure
of the net-clear command is not at all fatal.