Refactor tunnel/netif-mgmt, differentiate UP and RUNNING, etc.
This commit moves the code related to address/route/linkstate
management from `src/util/tunnel.[ch]` to the new
`src/util/netif-mgmt.[ch]`. This abstraction is a better fit. This
change also updates TunnelIPv6Interface accordingly.

One of the key changes in this commit is keeping track of the network
interface link state. Doing `ifconfig wpan0 down` will now gracefully
bring down the network interface. Likewise, doing `ifconfig wpan0 up`
will bring it back online. (This isn't supported on macOS yet)

We now differentiate the `UP` and `RUNNING` flags internally, however
the TUN driver on Linux does not make such a differentiation.
Nonetheless, maintaining this differentiation may be useful in the
future if we end up moving away from TUN.

The idea is that the `UP` flag is administratively controlled, whereas
`RUNNING` is controlled by the driver. For example, the user can call
`ifconfig eth0 up` to bring up their ethernet port, but if there is no
cable plugged into it the interface won't have the `RUNNING` flag, and
thus still isn't available. Likewise, the NCP may or may not be
provisioned for a network. Internally, we treat this as another way to
control the autoconnect property. If the interface is down,
autoconnect is off. If the interface is up, autoconnect will be on.