Add support for MLE discover scan
This commit includes the following changes: It updates the `TaskScan`
implementation to add support for the discovery scan in addition
energy and network scan operations It also adds a new dbus method for
MLE discovery scan. The wpanctl "scan" command is  updated to add
options for the new discover scan ("scan -d" to start a discovery
scan).