Add support for controlling steering data when joinable
This commit adds a new feature to allow setting of the steering data
when the joinability status gets changed (through `permit_join()`).
This feature requires the NCP `SPINEL_CAP_OOB_STEERING_DATA`
capability.

This commit defines two new wpan properties:

1) `OpenThread:SteeringData:SetWhenJoinable` (boolean) which defines
whether or not when the device is joinable the steering data is set.
The default setting is `false`. It is recommended that this property
is changed using `wpantund.conf` during initialization.

2) `OpenThread:SteeringData:Address` (eui64 value) which specifies the
address used to set the steering data (out of band) using the spinel
property `SPINEL_PROP_THREAD_STEERING_DATA` (the given eui64 is used
to compute the bloom filter which is used as the steering data).
Default value is all `0xff`s which means set the steering data/bloom
filter to accept/allow all.