Ensure to user a random xpanid during "form" if none was set explicitly earlier
This commits adds new flag `mXPANIDWasExplicitlySet` to track if the
XPANID is set explicitly after a NCP reset. This is then used to
ensure that a random xpanid is generated and used during "form"
opeation if non was set earlier.

This change addresses an issue where newly formed network always end
up using the default OpenThread XPANID value of `0xDEAD00BEEF00CAFE`.