Initial framework for fuzz testing
This is the first step in implementing a comprehensive fuzz test target
for wpantund. At the moment, the target is only fuzzing the config
file parser, which is nice but ultimately not really what we are interested
in. The plan is to extend this fuzz target to cover a much wider set
of input processing.

This fuzz target is currently set up for use by [libFuzzer][1], but
may in the future be used by other fuzzing systems.

Additional fuzz targets will be set up for more low-level module fuzzing.

[1]: http://llvm.org/docs/LibFuzzer.html