build: Precompiled Header Support
This commit adds support for both GCC and Clang precompiled headers.
Precompiled headers are used automatically if GCC or Clang is detected,
but it may be explicitly disabled by passing the `--disable-pch` argument
to the configure script.

The strategy used is to compute a precompiled header for each target and
include it when compiling each file. This should significantly improve
build times.