IPv6 address (unicast/multicast) and prefix entries
This commit contains the following changes:

It adds support for getting/updating multicast IPv6 address table from
NCP using spinel property `SPINEL_PROP_IPV6_MULTICAST_ADDRESS_TABLE`.

It also changes the model for tracking of addresses/prefixes.
`NCPInstanceBase` class keeps track of unicast/multicast and on-mesh-
prefixes in different tables. When an entry is added, the origin of the
entry is also recorded. We have three possible origins: (1) thread
NCP, (2) primary (tunnel) interface, or (3) user through a direct
command. The `NCPInstanceBase` is notified about updates to
addresses/prefixes through a set of methods which are then used by
tunnel interface and NCP plug-in implementation. If there is a change
`NCPInstanceBase` ensures to update other entities with newly
added/removed entry (e.g., if an address is added on the primary
interface, it is then updated on the NCP and vice versa). The new model
ensures that entries added by user (or through the primary interface)
persist over NCP reset. It also disallows an address/prefix added by
user to be removed by NCP or other entities.

Another change in this commit is on how on-mesh-prefixes corresponding
to new addresses are added. When a unicast IPv6 address is added on
primary interface, a corresponding prefix is derived and applied to
the NCP. A new wpan property "Daemon:SetDefaultRouteForAutoAddedPrefix"
is added to control the flags set for prefixes derived from an added
unicast address. If set to `true` when a prefix is auto-added it will
include the default route flag, otherwise it is set with this flag
cleared. The default setting for this property is false. The process
of adding SLACC addresses based on prefixes is also now done by the
`NCPInstanceBase`.

The `TunnelIPv6Interface` class is also modified to add support for
joining/leaving multicast addresses on interface. Another change is on
how the addresses are updated on the interface. If the interface is
offline, the addresses added to `TunelIpV6Interface` (using its
methods) are kept in a table and when the interface is brought online
the addresses are added. The new implementation also waits to get a
confirmation (through netlink) that an address is added before removing
the table entry (this helps with situation where adding an address may
first remove it on interface thus triggering an extra removed callback
notification).