Managing off-mesh-routes and their mapping to interface routes
This commit adds support for managing off-mesh-routes from Thread
network. All off-mesh routes (aka external routes) from NCP or user-
added (e.g, using `add_external_route()`) are stored in a
`std::multimap` list `mOffMeshRoutes`. Property "Thread:OffMeshRoutes"
can be used to get the full list which contains information about the
origin of each entry along with preference level.

This commit also adds logic to manage the routes on wpan network
interface corresponding to the off-mesh routes from the network: A
route is added on wpan network interface if there is an off-mesh route
on at least one other device within network AND either the same route
is not added by the device itself, or if it is added, it has a lower
preference level.

The change in this commit ensures that user-added off-mesh routes are
restored on NCP in case of an NCP reset.