string-utils: Fix for one-byte buffer overflow in `parse_string_into_data()`
This change addresses a one-byte buffer overflow in the function which
parses a hex string into data, `parse_string_into_data()`. The buffer
overflow would occur when the input string consisted of a single valid
hexadecimal digit, and would result in the a byte with the valid of
that digit in the upper nibble (and a zero in the lower nibble) being
written to an invalid location in the stack.

https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=3396