Fix problems disabling interface without low power support
It was noticed that disabling a Spinel interface caused the interface to
repeatedly reinitialize the NCP over and over again, largely due to the
fact that the OpenThread NCP doesn't support the low power feature yet.

This change makes sure that the interface can be successfully disabled
without causing wpantund to enter a reset loop.

Also deprecated `kWPANTUNDStatus_FeatureNotSupported` in favor of
`kWPANTUNDStatus_FeatureNotImplemented`.

Test: manual
Change-Id: Ifeb21badf34e4dd457dd65b352089f26c20eb0c0