ipc-binder: Update to make form/join API behave as expected
Without this change, a join failure will cause the NCP to enter the
commissioning state, which is not what the Binder API expects. This
change adds a internal option called "UseModernBehavior" which the form
and join tasks use to determine what behavior to use.

Eventually, the old behavior will be removed entirely, but for now we
need to ensure backward compatability.

Bug: b/67784777
Test: manual
Change-Id: I57ccaf937931aa0a0422328608785bc61b6f2f30