Adding new property wpantund "Thread:RouterTable"
This commit adds a new wpantund property "Thread:RouteTable" which
provides the list of all routers within the Thread network. This is
mapped to spinel `SPINEL_PROP_THREAD_ROUTER_TABLE`. This commit also
updated the `SpinelNCPTaskGetNetworkTopology` class to add support
for parsing a single router entry or entire router table.