ncp-spinel: Fix for alignment problems with in6_addr.
On some platforms, the copy constructor of `in6_addr` assumes that it
will always be aligned on 4-byte boundaries, causing crashes when that
is not the case. The most immediate fix is to use `memcpy` instead of
the default copy constructor. We are already doing this in most
instances, but it looks like a case slipped through. This commit
addresses this.

Bug: b/69921267
Test: join and form on RPI3
Change-Id: Ide82af2e361a721af7ff76d480a9e6b89e1148f8