Speculative/Diagnostic fix for oss-fuzz-3565
[oss-fuzz-3565](https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=3565)
describes an unreproducable (but frequently occuring) global buffer
overflow discovered by fuzzing. The immediate cause of the overflow is
the failure of `fd_set_merge()` to respect the size boundaries of an
`fd_set`, which is limited to accounting to no more than `FD_SETSIZE`
file descriptors.

However, the fact that we are getting anywhere close to the
`FD_SETSIZE` file descriptor limit indicates that we may very well be
leaking file descriptors. As such, this commit adds a quick file
descriptor leak check, intended to help identify file descriptor leaks
more quickly.

Such a file descriptor leak is likely to only be a practical problem
for fuzzing efforts, since the normal instance of wpantund would only
ever initialize once.