Diagnostic updates for oss-fuzz-3565
[oss-fuzz-3565](https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=3565)
describes an unreproducable (but frequently occuring) global buffer
overflow discovered by fuzzing. The immediate cause of the overflow is
the failure of `fd_set_merge()` to respect the size boundaries of an
`fd_set`, which is limited to accounting to no more than `FD_SETSIZE`
file descriptors. That issue was fixed by #263.

However, the fact that we are getting anywhere close to the
`FD_SETSIZE` file descriptor limit indicates that we may very well be
leaking file descriptors. #263 added some checks to detect file descriptor
leaks, but this commit introduces better checking that should be able to
ferrit out the exact sorts of test cases that induce this bug.