Fix parsing of input arguments in wpanctl mfg command
This commit fixes issues in parsing of input argument to wpanctl `mfg`
command. It allows any number of arguments to be provided and
ensures that the command string sent to NCP is limited to a maximum
size. Earlier implementation inherently assumed that user would only
use commands with at most 3 arguments, and if more arguments were
provided it would incorrectly send an uninitialized char buffer as
the command string to NCP which could lead to an NCP crash.

This commit also cleans up the code, removing unused variable
declarations  and addressing code style inconsistencies.