Replace an unchecked strcpy with a better strncpy
Issue flagged by the clang security checker.  Actual impact to product
is minimal -- the string being used as a source comes from dbus.  If
it was malicious it would be either because the attacker has a daemon
on the device already or because she has somehow hacked the pairing
flow into passing around a wrong network prefix.  At that point, a
buffer overrun and a resulting stack corruption would be the least of
our issues.  Nonetheless, it is best practice to check lengths so
that's what we're doing here.