Allow NCP reset to restart the initialization protothread earlier
This commit makes a change in the NCP initialization protothread  code
in `vprocess_init()` where the `mDriverState` is changed to
`INITIALIZING` (from `INITIALIZING_WAITING_FOR_RESET`) earlier in the
code and before sending a command to get the NCP protocol version.
This ensures that any received resets from NCP cause the protothread
to be restarted immediately.

This change addresses a delay issue with `wpantund` starting, where if
NCP is reset (e.g., by `spi-hdlc-adapter` process starting) while
wpantund is performing the NCP initialization, a wait time delay equal
to command response timeout (5 seconds) may be required for
initialization to restart.