New wpan property get/set/insert/remove handler model (#341)
This commit changes how the property get/set/insert/remove operations
are implemented. In the new model a `std::map` is used to associate
a wpan property name (as string) with a handler function pointer.

The new model helps simplify the implementation of many properties by
reusing the common code pattern used for getting/setting properties.

In particular a set of helper `register_get_handler` methods are
provided  to cover simpler cases, e.g., wpantund properties associated
with a spinel property with simple spinel packing reply format, or a
`ReplyUnpacker` function for more complex format. Similarly, for
"property set", a list of helper `register_set_handler` methods are
added which cover properties associated with a spinel property with a
(single) simple format. The `register` functions allow the user to
specify if the property needs persistence (i.e., the new value to be
saved in the `mSettings` map and pushed back on NCP upon reset),
and/or optionally require a capability check.