Add feature to control auto-addition of SLAAC address by wpantund (#356)
This commit adds a new feature and a related configuration property
"Daemon:IPv6:AutoAddSLAACAddress". This boolean property indicates
whether wpantund would generate and add an SLAAC address when an an
on-mesh prefix with the SLAAC flag is added. The SLAAC address is
added only if there is no existing address with the same prefix. By
default this feature is enabled (i.e. SLAAC addresses are added). It
can be changed using the above property during run-time or from
`wpantund.conf` file.