Add RCP version to "wpanctl status" when available
This commit adds RCP version to output of "wpanctl status" command
when POSX_APP capability is present in the list of NCP capabilities.
To enable this, `SpinelNCPInstance` is changed to save the RCP
version. A new property "POSIXApp:RCPVersion:Cached" is added to
get the cached value. Note that to ensure `status` command is
quick (does not block) all reported properties in `status` output
are cached. The `vprocess_init()` protothread is also updated to
retrieve RCP version during initialization.