Parse an empty DBUs message as an empty ValueMap dictionary (#417)
This commit enhances `DBUSHelpers::value_map_from_dbus_iter()` to
allow it to parse an empty DBus iter into a message as an empty
`ValueMap` dictionary. This addresses the case where a DBus command
(e.g., "JoinerAttach") is issued with no parameters (no appended DBUs
dictionary) expecting default parameters to be used for the command.