Add option to expect reset on start (#423)
On some scenario, POWER_ON event is emitted after wpantund tries to get
NCP version, which was treated as unexpected reset and causing wpantund
to reset the NCP again. This PR adds an option to expect a reset on
start, so that NCP will not get reset multiple times on wpantund start.