Make "reset" wait till end of NCP init (#430)
This commit convert the `SpinelNCPControlInterface::reset()` API to
wait till end of NCP re-initilaization (instead of returning after
issuing the `RESET` command to NCP). This is realized by adding a
feature in `TaskSendCommand` to allow an optional "final check
handler" to be added to the command. If provided after the commands
are sent to NCP, the task will invoke the provided handler and waits
until the handler returns `true` (within the given timeout interval).