Feature to auto add on-mesh prefixes as interface route
This commit adds a new feature in `wpantund` to auto add a route on
the host primary interface corresponding to on-mesh prefixes from
Thread network.

This feature can be enabled through a newly added wpan property with
name "Daemon:OnMeshPrefix:AutoAddAsIntterfaceRoute" which is enabled
by default.

When enabled, wpantund would add a route on host primary interface for
any prefix from thread network (with on-mesh flag set). This in turn
ensures that traffic destined to an IPv6 address matching the prefix
would be correctly forwarded to the `wpan` interface on host.