Add support for generating PSKc from pass-phrase, net-name, and XPANId
This commit adds support in wpantund and wpanctl to generate PSKc
from a given commissioning pass-phrase, network name and XPANID.
The spinel property `SPINEL_PROP_MESHCOP_COMMISSIONER_GENERATE_PSKC`
is used for this purpose. A new DBUS API (v1) is added to expose
this functionality to other processes. wpanctl `commissioner`
command is updated to provide `gen-pskc`.