Relax parsing of set value on "Dataset:Active/PendingTimestamp" (#445)
This commit updates `any_to_uint64()` to allow a string representing
to be optionally treated as a number (decimal number or hex value
starting with `0x`). This change relaxes the input format when setting
"Dataset:Active/PendingTimestamp" properties, while keeping the parsing
of XPANID as before (expecting it to be a 16 char hex value).